Kill a running puppet using puppet

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -2,5 +2,5 @@
 
 exec { 'kill_process':
   command => 'kill 17818',
-  onlyif  'ps -p 17818',
+  onlyif  => 'ps -p 17818',
 }
